refactor: split MCP server.go into 34 domain files
Agent-Task: mcp-server-split
Agent-Model: SOLO
Agent-Decision: Split 2510-line server.go into handler, register, types, helpers files

Before: server.go 2510 lines (types + 63 tool registrations + 69 handlers + helpers)
After:  server.go 71 lines (entry point only)

New files:
- types.go (207 lines): interfaces (Deployer, PreflightErrorInfo) + 12 structs
- helpers.go (80 lines): validateVolumePath, validateImageRegistry, etc.
- 16 handler_*.go files (1420 lines): tool handler implementations
- 16 register_*.go files (870 lines): tool registration functions

NewServer() now delegates to 16 register*() functions by domain.
contextManager global moved to context_manager.go.
